Edit sensor functions

+++ include/sensor_info.h
@@ -8,15 +8,15 @@ private:
     int sensor_id_;
     std::string type_;
     std::string unit_;
-    double (*function_) ();
+    float (*function_) ();
 
 
 public:
 
-    SensorInfo(int sensor_id, std::string type, std::string unit, double (*function) ());
+    SensorInfo(int sensor_id, std::string type, std::string unit, float (*function) ());
 
     int GetSensorId();
     std::string GetType();
     std::string GetUnit();
-    double GetValue();
+    float GetValue();
 };
